Moving "is_<logtype>_start_line" utility functions to joutstructure as its only foreseeable utility is in its initializing a joutstructure object. "is_lowdin_start_line" not included due to usage in other class initializations. Moving "get_joutstructures_start_idx" and "get_joutstructure_step_bounds" to joutructures module.
