+++ src/pymatgen/io/jdftx/utils.py
@@ -252,7 +252,7 @@ def _brkt_list_of_3x3_to_nparray(lines: list[str], i_start: int = 0) -> np.ndarr
     return out
 
 
-elec_min_start_flag: str = "-------- Electronic minimization -----------"
+# elec_min_start_flag: str = "-------- Electronic minimization -----------"
 
 
 def get_colon_var_t1(linetext: str, lkey: str) -> float | None:
@@ -274,66 +274,10 @@ def get_colon_var_t1(linetext: str, lkey: str) -> float | None:
     return colon_var
 
 
-def is_strain_start_line(line_text: str) -> bool:
-    """Return True if the line_text is the start of strain log message.
-
-    Return True if the line_text is the start of a log message for a JDFTx
-    optimization step.
-
-    Parameters
-    ----------
-    line_text: str
-        A line of text from a JDFTx out file
-
-    Returns
-    -------
-    is_line: bool
-        True if the line_text is the start of a log message for a JDFTx
-        optimization step
-    """
-    return "# Strain tensor in" in line_text
-
-
-def is_lattice_start_line(line_text: str) -> bool:
-    """Return True if the line_text is the start of lattice log message.
-
-    Return True if the line_text is the start of a log message for a JDFTx
-    optimization step.
-
-    Parameters
-    ----------
-    line_text: str
-        A line of text from a JDFTx out file
-
-    Returns
-    -------
-    is_line: bool
-        True if the line_text is the start of a log message for a JDFTx
-        optimization step
-    """
-    return "# Lattice vectors:" in line_text
-
-
-def is_forces_start_line(line_text: str) -> bool:
-    """Return True if the line_text is the start of forces log message.
-
-    Return True if the line_text is the start of a log message for a JDFTx
-    optimization step.
-
-    Parameters
-    ----------
-    line_text: str
-        A line of text from a JDFTx out file
-
-    Returns
-    -------
-    is_line: bool
-        True if the line_text is the start of a log message for a JDFTx
-        optimization step
-    """
-    return "# Forces in" in line_text
-
-
+# This function matches the format of the generic "is_<x>_start_line" functions specific to
+# the JOutStructure object initialization, but is not moved to the joutstructure module
+# as it is also used in methods for other JDFTx IO modules (e.g. JOutStructures) so it
+# is kept here to avoid circular imports.
 def is_lowdin_start_line(line_text: str) -> bool:
     """Return True if the line_text is the start of Lowdin log message.
 
@@ -352,72 +296,6 @@ def is_lowdin_start_line(line_text: str) -> bool:
         JDFTx out file
     """
     return "#--- Lowdin population analysis ---" in line_text
-
-
-def get_joutstructure_step_bounds(
-    out_slice: list[str],
-    out_slice_start_flag: str = elec_min_start_flag,
-) -> list[list[int]]:
-    """Return list of boundary indices for each structure in out_slice.
-
-    Return a list of lists of integers where each sublist contains the start and end
-    of an individual optimization step (or SCF cycle if no optimization).
-
-    Parameters
-    ----------
-    out_slice: list[str]
-        A slice of a JDFTx out file (individual call of JDFTx)
-
-    Returns
-    -------
-    bounds_list: list[list[int, int]]
-        A list of lists of integers where each sublist contains the start and end
-        of an individual optimization step (or SCF cycle if no optimization)
-    """
-    bounds_list = []
-    bounds = None
-    end_started = False
-    for i, line in enumerate(out_slice):
-        if not end_started:
-            if out_slice_start_flag in line:
-                bounds = [i]
-            elif (bounds is not None) and (is_lowdin_start_line(line)):
-                end_started = True
-        elif not len(line.strip()) and bounds is not None:
-            bounds.append(i)
-            bounds_list.append(bounds)
-            bounds = None
-            end_started = False
-            # else:
-            #     warnmsg = f"Line {i-1} ({out_slice[i-1]}) triggered \
-            #         end_started, but following line is empty. Final step_bounds \
-            #             may be incorrect. "
-            #     warnings.warn(warnmsg, stacklevel=2)
-    return bounds_list
-
-
-def get_joutstructures_start_idx(
-    out_slice: list[str],
-    out_slice_start_flag: str = elec_min_start_flag,
-) -> int | None:
-    """Return index of first line of first structure.
-
-    Return the index of the first line of the first structure in the out_slice.
-
-    Parameters
-    ----------
-    out_slice: list[str]
-        A slice of a JDFTx out file (individual call of JDFTx)
-
-    Returns
-    -------
-    i: int
-        The index of the first line of the first structure in the out_slice
-    """
-    for i, line in enumerate(out_slice):
-        if out_slice_start_flag in line:
-            return i
-    return None
 
 
 def correct_geom_opt_type(opt_type: str | None) -> str | None:
@@ -443,106 +321,6 @@ def correct_geom_opt_type(opt_type: str | None) -> str | None:
         else:
             opt_type = None
     return opt_type
-
-
-def is_magnetic_moments_line(line_text: str) -> bool:
-    """Return True if the line_text is start of moments log message.
-
-    Return True if the line_text is a line of text from a JDFTx out file
-    corresponding to a Lowdin population analysis.
-
-    Parameters
-    ----------
-    line_text: str
-        A line of text from a JDFTx out file
-
-    Returns
-    -------
-    is_line: bool
-        True if the line_text is a line of text from a JDFTx out file
-        corresponding to a Lowdin population
-    """
-    return "magnetic-moments" in line_text
-
-
-def is_charges_line(line_text: str) -> bool:
-    """Return True if the line_text is start of charges log message.
-
-    Return True if the line_text is a line of text from a JDFTx out file
-    corresponding to a Lowdin population analysis.
-
-    Parameters
-    ----------
-    line_text: str
-        A line of text from a JDFTx out file
-
-    Returns
-    -------
-    is_line: bool
-        True if the line_text is a line of text from a JDFTx out file
-        corresponding to a Lowdin population
-    """
-    return "oxidation-state" in line_text
-
-
-def is_ecomp_start_line(line_text: str) -> bool:
-    """Return True if the line_text is the start of ecomp log message.
-
-    Return True if the line_text is the start of a log message for a JDFTx
-    optimization step.
-
-    Parameters
-    ----------
-    line_text: str
-        A line of text from a JDFTx out file
-
-    Returns
-    -------
-    is_line: bool
-        True if the line_text is the start of a log message for a JDFTx
-        optimization step
-    """
-    return "# Energy components" in line_text
-
-
-def is_posns_start_line(line_text: str) -> bool:
-    """Return True if the line_text is the start of posns log message.
-
-    Return True if the line_text is the start of a log message for a JDFTx
-    optimization step.
-
-    Parameters
-    ----------
-    line_text: str
-        A line of text from a JDFTx out file containing the positions of atoms
-
-    Returns
-    -------
-        is_line: bool
-            True if the line_text is the start of a log message for a JDFTx
-            optimization step
-    """
-    return "# Ionic positions" in line_text
-
-
-def is_stress_start_line(line_text: str) -> bool:
-    """Return True if the line_text is the start of stress log message.
-
-    Return True if the line_text is the start of a log message for a JDFTx
-    optimization step.
-
-    Parameters
-    ----------
-    line_text: str
-        A line of text from a JDFTx out file
-
-    Returns
-    -------
-    is_line: bool
-        True if the line_text is the start of a log message for a JDFTx
-        optimization step
-    """
-    return "# Stress tensor in" in line_text
 
 
 ############################################
