+++ tests/io/jdftx/test_utils.py
@@ -2,13 +2,15 @@ from __future__ import annotations
 
 import pytest
 
+from pymatgen.io.jdftx.joutstructures import get_joutstructures_start_idx
+
 # from pymatgen.io.jdftx.generic_tags import flatten_list
 # from pymatgen.io.jdftx.jdftxoutfileslice_helpers import find_first_range_key, get_start_lines
 # from pymatgen.io.jdftx.joutstructure_helpers import get_joutstructures_start_idx
 from pymatgen.io.jdftx.utils import (
     find_first_range_key,
     flatten_list,
-    get_joutstructures_start_idx,
+    # get_joutstructures_start_idx,
     get_start_lines,
     multi_getattr,
     multi_hasattr,
